requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data acquisition
courtvisionpython>=0.1.0
requests>=2.31.0
aiohttp>=3.9.0

# Visualization & reporting
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
